src/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "AI Applications Chassis - AI Apps Best Practices Tailored for Azure - Backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi ~=0.115.6",
    "uvicorn ~=0.32.1",
    "python-dotenv ~=1.0.1",
    "pyyaml>=6.0.2",
    "openai>=1.61.0",
    "semantic-kernel>=1.22.1",
    "opentelemetry-exporter-otlp-proto-grpc>=1.29.0",
    "opentelemetry-instrumentation>=0.50b0",
    "azure-monitor-opentelemetry-exporter>=1.0.0b34",
    "azure-identity>=1.19.0",
    "azure-ai-inference[opentelemetry]>=1.0.0b9",
    "azure-search-documents"
    
]
